Feature/job services

+++ src/main/java/com/yusuf/online/order/system/core/service/SimpleProductService.java
@@ -3,7 +3,6 @@ package com.yusuf.online.order.system.core.service;
 import com.yusuf.online.order.system.core.entity.Product;
 import com.yusuf.online.order.system.core.mapper.ProductMapper;
 import com.yusuf.online.order.system.core.model.dto.ProductDTO;
-import com.yusuf.online.order.system.core.model.dto.UserDTO;
 import com.yusuf.online.order.system.core.model.request.ProductListRequest;
 import com.yusuf.online.order.system.core.repository.ProductRepository;
 import com.yusuf.online.order.system.core.service.base.ProductService;
@@ -11,6 +10,7 @@ import com.yusuf.online.order.system.core.service.base.UserService;
 import jakarta.persistence.EntityExistsException;
 import jakarta.persistence.EntityNotFoundException;
 import java.util.List;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -26,13 +26,12 @@ public class SimpleProductService implements ProductService {
 
   @Override
   public ProductDTO save(ProductDTO productDTO) {
-    final String productName = productDTO.getName();
-    if (repository.existsByName(productName)) {
-      throw new EntityExistsException(
-          String.format("%s isimli ürün zaten mevcut. Sadece güncelleme yapabilirsiniz",
-              productName));
-    }
     final Integer currentUserId = userService.getCurruntUser().getId();
+    if (repository.existsByNameAndSellerId(productDTO.getName(), currentUserId)) {
+      throw new EntityExistsException(
+          String.format("%s ürünü için ürün kaydı mevcuttur. Yeni kayıt atılamaz.",
+              productDTO.getName()));
+    }
     final Product product = productMapper.convertToEntity(productDTO);
     product.setSellerId(currentUserId);
     final Product savedProduct = repository.save(product);
@@ -41,39 +40,46 @@ public class SimpleProductService implements ProductService {
 
   @Override
   public ProductDTO update(ProductDTO productDTO) {
-    final UserDTO currentUser = userService.getCurruntUser();
-    final String productName = productDTO.getName();
-    Product product = repository.findByNameAndSellerId(productName, currentUser.getId())
-        .orElseThrow(() -> new EntityNotFoundException(
-            String.format("%s'ye ait %s ürünü bulunamadığından güncelleme yapılamadı.",
-                currentUser.getEmail(), productName)));
-    productMapper.updateEntity(productDTO, product);
+    final Integer currentUserId = userService.getCurruntUser().getId();
+    final Product product = repository.findById(productDTO.getId()).map(p -> {
+      if (!Objects.equals(currentUserId, p.getSellerId())) {
+        throw new IllegalArgumentException(
+            "Güncellemek istenilen ürün size ait değil. Herkes sadece kendine ait ürünü güncelleyebilir.");
+      }
+      productMapper.updateEntity(productDTO, p);
+      return p;
+    }).orElseThrow(() -> new EntityNotFoundException(
+        String.format("%s ID'ye ait ürün kaydı bulunamadığından güncelleme yapılamadı.",
+            productDTO.getId())));
     final Product savedProduct = repository.save(product);
     return productMapper.convertToDTO(savedProduct);
   }
 
   @Override
-  public void deleteByProductId(String productName) {
-    final UserDTO currentUser = userService.getCurruntUser();
-    final Product product = repository.findByNameAndSellerId(productName, currentUser.getId())
+  public void deleteByProductId(Integer id) {
+    final Integer currentUserId = userService.getCurruntUser().getId();
+    final Product product = repository.findById(id)
         .orElseThrow(() -> new EntityNotFoundException(
-            String.format("%s 'e ait %s ürünü bulunamadığından silme işlemi gerçekleştirilemedi.",
-                currentUser.getEmail(),
-                productName)));
+            String.format(
+                "%s ID'ye ait ürün kaydı bulunamadığından silme işlemi gerçekleştirilemedi.", id)));
+    if (!Objects.equals(currentUserId, product.getSellerId())) {
+      throw new IllegalArgumentException(
+          "Silmek istenilen ürün size ait değil. Herkes sadece kendine ait ürünü silebilir.");
+    }
     repository.delete(product);
   }
 
   @Override
   public List<ProductDTO> getProductsByNameAndDescription(ProductListRequest request) {
     final List<Product> productsByNameAndDescription = repository.getProductsByNameAndDescription(
-        request.getProductName(), request.getDescription());
+        request.getName(), request.getDescription());
     return productMapper.convertAllToDTO(productsByNameAndDescription);
   }
 
   @Override
   public ProductDTO getProductById(Integer id) {
     final Product product = repository.findById(id).orElseThrow(
-        () -> new EntityNotFoundException(String.format("%s ID'ye ait ürün bulunamadı!")));
+        () -> new EntityNotFoundException(String.format("%s ID'ye ait ürün bulunamadı!",id)));
     return productMapper.convertToDTO(product);
   }
